refactor: move repository CTA to sidebar
